requirements.txt
opencv-python-headless
atproto
schedule
Pillow
